UPD: reconstruct route module
