+++ unfazed/type/route.py
@@ -1,5 +1,8 @@
 import typing as t
 
+from pydantic import BaseModel
+
 type HttpMethod = t.Literal[
     "GET", "POST", "PATCH", "DELETE", "HEAD", "OPTIONS", "PUT", "TRACE"
 ]
+SUPPOTED_REQUEST_TYPE = t.Union[str, int, float, t.List, BaseModel]
